+++ packages/amplify-sdk/src/auth.js
@@ -48,7 +48,7 @@ export default class Auth {
 	 * @param {String} [opts.baseUrl] - The base URL to use for all outgoing requests.
 	 * @param {String} [opts.clientId] - The client id to specify when authenticating.
 	 * @param {String} [opts.clientSecret] - The secret token to use to authenticate.
-	 * @param {String} [opts.env=prod] - The environment name. Must be `dev`, `preprod`, or `prod`.
+	 * @param {String} [opts.env=prod] - The environment name. Must be `staging` or `prod`.
 	 * The environment is a shorthand way of specifying a Axway default base URL.
 	 * @param {Function} [opts.got] - A reference to a `got` HTTP client. If not defined, the
 	 * default `got` instance will be used.
@@ -395,7 +395,7 @@ export default class Auth {
 	 * @param {String} [opts.baseUrl] - The base URL to use for all outgoing requests.
 	 * @param {String} [opts.clientId] - The client id to specify when authenticating.
 	 * @param {String} [opts.code] - The authentication code from a successful interactive login.
-	 * @param {String} [opts.env=prod] - The environment name. Must be `dev`, `preprod`, or `prod`.
+	 * @param {String} [opts.env=prod] - The environment name. Must be `staging` or `prod`.
 	 * The environment is a shorthand way of specifying a Axway default base URL.
 	 * @param {Boolean} [opts.manual=false] - When `true`, it will return the auth URL instead of
 	 * launching the auth URL in the default browser.
@@ -474,7 +474,7 @@ export default class Auth {
 	 *
 	 * @param {Object} [opts] - Various options.
 	 * @param {String} [opts.baseUrl] - The base URL to use for all outgoing requests.
-	 * @param {String} [opts.env=prod] - The environment name. Must be `dev`, `preprod`, or `prod`.
+	 * @param {String} [opts.env=prod] - The environment name. Must be `staging` or `prod`.
 	 * The environment is a shorthand way of specifying a Axway default base URL.
 	 * @param {String} [opts.realm] - The name of the realm to authenticate with.
 	 * @param {String} [opts.url] - An optional URL to discover the available endpoints.