+++ packages/amplify-utils/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.0.6 (Feb 2, 2022)
+
+ * chore: Updated dependencies.
+
 # v1.0.5 (Jan 14, 2022)
 
  * chore: Updated dependencies.